+++ R/defaultControl.R
@@ -30,9 +30,9 @@ defaultControl<-function() {
     Catm = 386,
     ndailysteps = 24,
     thermalCapacityLAI = 1000000,
-    defaultWindSpeed = 2.5, #m/s
+    defaultWindSpeed = 2.5 #m/s
     # growth
-    storagePool = "none"
+    # storagePool = "none"
     
 #     #For water balance
     #For forest dynamics