+++ NEWS.md
@@ -5,6 +5,7 @@
 # Version 0.10.0
 - Reorganization of growth function
 - Clarification of gross and net photosynthesis
+- 'spwb_resetInputs' to 'resetInputs'
 
 # Version 0.9.1
 - Small bug fixes